Workbench Junk
Added two things from workbench. To separate them from everything else and to make it easier to find, I made a workbench tag. Workbench construction differs from the other stuff so I gave it it's own area. That can be changed if necessary. I may or may not finish this but I have Heavy armor finished and armor plates. From there spears should be pretty easy to do and so should Arrows now that I know how they are made.